Add initial support for PGO instrumentation/use in inputgen
This patch adds flags (and a test) so that the main input-gen binary can
add instrumentation to modules and also compile the run modules with a
PGO profile. This enables downstream uses of PGO, such as block coverage
analysis. Passing the flags directly into the input-gen binary for
recompiling with a PGO profile probably isn't what will be used for
large scale testing (instead just opting to recompile the bitcode), but
probably won't hurt to support.

This is also currently not selective in terms of what gets instrumented.
This means that parts of the runtime will end up in the profile. This
shouldn't be an issue, but something that we may want to work on in the
future.